+++ apps/backoffice-v2/src/common/components/organisms/TextEditor/components/ShortcutKey.tsx
@@ -22,7 +22,7 @@ export const ShortcutKey = React.forwardRef<HTMLSpanElement, ShortcutKeyProps>(
           <kbd
             key={shortcut.symbol}
             className={ctw(
-              'inline-block min-w-2.5 text-center align-baseline font-sans text-xs font-medium capitalize text-[rgb(156,157,160)]',
+              'min-w-2.5 inline-block text-center align-baseline font-sans text-xs font-medium capitalize text-[rgb(156,157,160)]',
 
               className,
             )}